Json dimmer values instead of static html
Now using dimmer values received from the JSON call instead of static
HTML values. I'm not able to test the custom levels received from the
JSON call myself but tested it by changing the values at runtime in the
debugger. Can someone verify if it is working correctly in a real case
scenarios?