Eco devices, further enhancements and switch to standard helper functions (#1268)
* First attempt at getting rid of P1 Smart Meter dependency and use standard helpers to send data.

New helpers need to be added for water and gas meters as they don't seem to exist.

* First attempt at getting rid of P1 Smart Meter dependency and use standard helpers to send data.

New helpers need to be added for water and gas meters as they don't seem to exist.

* Bug fix that might create a useless device when  subscription is HP/HC

* Devices are now using standard helpers rather than sDecodeRXMessage

New icon for water meter added

* Revert some files to the original version as changes are not needed anymore

* Revert some files to the original version as changes are not needed anymore

* All references and dependencies to P1Smartmeter suppressed
sDecodeRX function not used anymore. Switch to the standard helpers Sendxxxxxx
EJP, Tempo, Basic and Heures Creuses contracts now managed by EcoDevice code

* - All calls to sDecodeRX suppressed and replaced with standard helpers (SendKwhMeter, SendMeterSensor....)
- All 4 main French electricity rates now supported (Basic, Tempo, EJP and "Heures Creuses")
- New functionality: a message device returns the current rate for multiple rates contracts (EJP, HC and Tempo)
- Pulse counters now rely on SendMeterSensor which does a much better job than is P1SmartMeter based counterpart
     when tracking water or gaz consumption.

The EcoDevices device and the Teleinfo protocol are specifically French, I coded all messages and default
names in French as is is very unlikely that anybody non speaking French would ever read them. Is that OK?

* - All calls to sDecodeRX suppressed and replaced with standard helpers (SendKwhMeter, SendMeterSensor....)
- All 4 main French electricity rates now supported (Basic, Tempo, EJP and "Heures Creuses")
- New functionality: a message device returns the current rate for multiple rates contracts (EJP, HC and Tempo)
- Pulse counters now rely on SendMeterSensor which does a much better job than is P1SmartMeter based counterpart
     when tracking water or gaz consumption.

The EcoDevices device and the Teleinfo protocol are specifically French, I coded all messages and default
names in French as is is very unlikely that anybody non speaking French would ever read them. Is that OK?

* Added translation to comments in EcoDevices.h

Fix include that may be the cause of compile failures when pulling requests

* Added translation to comments in EcoDevices.h

Fix include that may be the cause of compile failures when pulling requests