OpenWebNet: bug fix and burglar alarm battery, network and status
Fix wrong automatic creation of auxiliary device and update status of
dry contact, introduced the development of CEN commands.

Unit must be left free for the interface, so the alarm zone is used as
the 'where' and not as 'interface' in UpdateAlarm.
After this update, the alarm device will have a different ID.
Adding devices to the battery status, network status and alarm
maintenance.