Create custom.css
I've added a new theme.

Making the theme was quite an adventure. The HTML/CSS for Domoticz is extremely messy and doesn't follow standards.

It would be great to clean and update the HTML front-end sometime.. I think it really matters to the succes of Domoticz, which currently has some really great features that others don't have. But they are catching up, and they do look pretty and are more usable.

Usability matters. A lot!