Related to https://github.com/domoticz/domoticz/pull/1162
This commit includes szczukot's changes, plus suggestions by gizmocuz.

Now sensors are not updated in bursts, but one at a time on a fixed intervals.

BIG CHANGE: All sensor updates are now done with SendTemperature() and similar functions.
This will create new temperature and humidity sensors, because the subtype changes..
Perhaps there should be a script that automatically updates the database?

I only have 1 wire temperature sensors, so these changes are not tested on other devices.