Added CSS classes
H2 should not be inside P tags, so I did remove a P and a BR there. H2 is a block-style element, so a BR afterwards is unnecessary (will break automatically).
Also let javascript add relevant tags to the body, like the name of the page, mobile preference, and whether the user is admin or not. This is very useful for styling.