Change status color via CSS class
Added new statusClass variable to scope of temperature and weather widgets controller'. These are some of the few widgets still heavily reliant on Angular.
This will add a status class to the item div in the corresponding templates, which will give a lot of freedom to change the display of items depending on their status (protected, timeout, etc).