Custom JS load
The index.html now loads a custom.js file form the theme folder. This gives a lot of possibilities to themers, like adding data visualisations, incorporating other data sources, etc etc.

The file also has 10 simple DIV's all the way at the bottom of the screen, which can be used to communicate settings form CSS to JS. This is done by setting the Z-index of those DIV's as a signal that a specific CSS file has been incorporated.