added a container where themesettings are loaded into
jquery loads a themesettings.html file into a new div with the id "themesettingscontainer".

Technically speaking this theme settings div now has a different depth than the other items on the page, because that container is added. But it practise it turns out there is no CSS that was targetting that level specifically, so everything is fine. Lucky!