improved version
Now with 9 options, although 3 of those are not currently used in the default.

It would be great if Domoticz could allow for 9 generic string variables to be set. So they don't all have to be boolean, it would be great if they could just be strings. That way a theme could in theory allow a user to set a number (width, number of columns, etc) for a theme.