Revert "bugfix: return 'nvalue' iso 'svalue' for humidity levels in datapush handling"
While this fixes the issue with dummy devices it breaks real devices because of inconsistancies in data storage and data parsing in the datapush module between dummy devices and real devices.
see issue #1656 for details on the issue and a possible fix for both situations.
Please revert this fix as it breaks more than it fixes.

This reverts commit f24fa115bff3cd6948b2ad306ed14e35f6645483.