Allow scripts to set Evohome setpoints
... and clean up the polymorphism in MainWorker:SetSetPoint() somewhat. The
three-argument version of SetSetPoint() only worked for Evohome devices,
while the one-argument version worked for everything, by calling through
to the Evohome version where necessary. I'm not convinced I would have
given those functions the same name, but OK. If the intent is that this
interface will be used to handle Domoticz-based scheduling for currently
"dumb" devices at some point in the future, that makes some sense.

Fix up the three-argument version to call through to SetSetPointInt() for
devices that it doesn't support, and then change the script invocations
so that they can handle up to three arguments.

Also sift the string matching on PermanentOverride/TemporaryOverride into
SetSetPoint() from the web server code, to avoid duplicating it.

Now with a little bit of scripting and a virtual switch, I can say
'Alexa, turn on bathtime" and ths kids' radiators all turn up to 23°C
for the next hour.