Report timer dates correctly as yyyy-mm-dd
This actually changes the JSON returned on the wire to clients, which means
it warrants additional review because it might potentially affect third-party
clients which talk to Domoticz.

I think it's fine though — nobody in their right mind would *ever* see a
local dialect on the wire and *not* assume that it's going to change and
that they're going to have to adapt.