OnkyoAVTCP: Improve support for models without NRIQSTN
Learn the available values of the input selectors as they appear, mapping
them to the known names where we can. Also create logical devices as they
appear.

This means that users of pre-NRIQSTN devices should be able to make
Domoticz learn the available options purely by iterating over the zones,
turning stuff on and off and selecting all the inputs.