Icon for "Mobile Devices" in the menu
Both an "on" version and a "off" version of a mobile phone icon.

Based on a public domain image.
https://openclipart.org/detail/19480/modern-touch-phone-mobile
- added little shadow and border spacing, conforming to the style of all other icons I've created for Domoticz.
- created an 'off' layer

In future It might be nice to use CSS Filters to greyscale icons when they are "off". That way only half the icons would need to be shipped. This has a number of advantages.
- Less things to download means a faster and more responsive interface, especially on mobile devices.
- Easier maintenance
- Less controller code required to switch icons. Just toggle an "on" or "off" class to the icon.