Only require libudev if USE_STATIC_OPENZWAVE is set
When we have a dynamic libopenzwave.so, we don't need to explicitly link
against libudev — and *shouldn't* link against libudev, because some
systems (like OpenWRT) don't have it and build OpenZWave not to use it.

It's only needed when we're linking against a static libopenzwave.a,
which doesn't encode the other libraries that it depends on. (Actually
this *ought* to work too via pkg-config but OpenZWave doesn't correctly
put them into its Libs.private, and we support building against OpenZWave
from its build dir without using pkg-config anyway.)