The devices are not created correctly
The function PushAndWaitRx must be call after the safe_query in order to have the if (result.empty(s) statement working correctly !

That is a not a great code, but this is how it was originally implemented. the latest changes which have been made in 0af09b28e215e31d726081e07ea4a63176919bbe have introduce the issue.