dzVents: Avoid rescanning script directories (#2998)
Rather than rescanning repeatedly to build the handler lists for each
event, scan just once and retain the full list for later processing.

Eventually, we might even be able to make Domoticz store this for itself
across multiple invocations of dzVents.lua. But for now, not doing it
multiple times in a single invocation is a good start.