Fix ui-popup li height in elemental/custom.css
The default line height of 4px in the elemental theme is squashing the ui-popup used for the evohome popup menu.
It looks bad as the icons are clipped and the selection is difficult to use on a touch device.
Restore the original height purely for li child elements of ui-popup class.