Fix multiple crashing bugs in plugin HTTP protocol handler
CPluginProtocolHTTP::ProcessInbound assumes in many place that tokens are
completely contained within a received buffer.  If that is not the case
then the code may read beyond the end of the buffer, or believe it has
received the full response when actually there is still more data to come.
Depending on the circumstances this will either cause the plugin framework
to crash immediately, or on the next attempt next to parse and collate
an imcoming message.