Increase camera fetch timeout to 10s
I use a CGI script to fetch a snapshot from an RTSP stream:

 #!/bin/sh
 cat <<EOF
 Content-Type: image/jpeg

 EOF
 ffmpeg -skip_frame nokey -i rtsp://${CAMERA}:554/ch01.264 \
      -qscale:v 4 -frames:v 1 -f image2pipe -

It takes a little over 5 seconds, every time. I had to bump the
timeout up to 10s. This could be configurable, maybe, and perhaps
we should even teach Domoticz to spawn ffmpeg for itself for an
RTSP camera URI. But a ten-second timeout, and doing the image
processing on a more powerful box, works well in the short term.