Prevent use of wrong commands in switchcmd parm of switchscene JSON
Code logic cause wrong spelled commands to be interpreted as Off
So if a user misspelled On like on it would cause an unexpected
behavior.

It could well be that the way I tried to fix this is not conform the
standard domoticz way of handling it.
In that case please advise how to tackle this.