Cope with open-zwave-read-only being under the domoticz directory
I had tested with it being in ../open-zwave-read-only but the Travis build
actually checks it out as a subdirectory of the domoticz checkout.

Cmake still finds the library because we set CMAKE_LIBRARY_PATH explicitly
and our hints about ../open-zwave-read-only in CMakeLists.txt are not
needed.

Find the include files in the cpp/src directory based on where the
library was found, instead of just hard-coding the same hints.