unzip: reduce file name size to 65535 to work with external minizip
The external minizip project has changed the unzGetCurrentFileInfo()
function to take a uint16_t as the filename size, instead of a uLong
as in the original version in zlib.

(Reported as https://github.com/nmoinvaz/minizip/issues/490 but it
was 3½ years ago and might be too late to fix it now, although changing
it back to a *larger* type is a lot safer than reducing the size, and
perhaps they should.)

This means that our 65536-byte buffer gets truncated to zero, as the
compiler tells us when we build agaisnt the external minizip:

domoticz/main/unzip_stream.h:140:50: warning: conversion from ‘long unsigned int’ to ‘uint16_t’ {aka ‘short unsigned int’} changes value from ‘65536’ to ‘0’ [-Woverflow]
  140 |     unzGetCurrentFileInfo(handler_, &info, path, sizeof(path), NULL, 0, NULL, 0);
      |                                                  ^~~~~~~~~~~~

Reduce the buffer size to 65535 bytes instead.