Pressure: allow 2 decimals for pressure sensors
Plugwise Adam reports boiler pressure in 2 decimals, e.g. 1.53 bar, but
domoticz internally lops off the second decimal. To make it worse,
SQLHelper them drops *all* decimals.

In the context of Plugwise, the difference between '2.49' and '2.00' is
significant, presenting and storing both as '2' is very misleading. I
noticed this when wanting to use Domoticz to see if the repair to the
leak in the central heating was successful.

Increase precision by one decimal everywhere: "%.2f", except for
SQLHelper, that changes to "%.1f".

The net result is that data push to e.g. InfluxDB will have actually
useful values.

Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>