explicitly set C++ version to 14
This was already the case with MSVC which defaults to C++14. In the
case of clang, this was causing build failures as C++14 features are
used in the codebase. In the case of gcc, the libc++ used there is
non-standards conformant and exposes C++14 features to C++11.

This fixes compilation under FreeBSD which uses clang and libcxx as the
codebase introduced C++14 stuff recently.

Signed-off-by: Rosen Penev <rosenp@gmail.com>