fix some lambdas
modernize-avoid-bind requires C++14 to run. Unfortunately, manually
editing the lambdas to be compatible with C++11 was not a smart thing
to do. Specifically, it resulted in quite significant deviations from
the clang-tidy transformations.

Just in case, make the lambdas more similar to the ones generated by
clang-tidy.

Also fix bad lambda in tcpproxy_server.cpp . handle_downstream_write
was switched with the read variant somehow...

Signed-off-by: Rosen Penev <rosenp@gmail.com>