Add API/JSON to delete shortlog and or calendarlog on a per idx base
detailed history for some devices might be interresting while for other devices it will never be looked at
this API call provides the option to selectively delete the shortlog and/or calendarlog of a device without the need
to use external tools.
Like it is already available for switchtype devices and text sensors.

/json.htm?idx=<idx>&type=command&param=deletehistory&historytype=calenderlog
/json.htm?idx=<idx>&type=command&param=deletehistory&historytype=shortlog