Update HEOS.cpp
Let all message start with 'HEOS'. Some log messages start with 'DENON by HEOS', others with 'HEOS by DENON'. The first is incorrect because HEOS is a system from Denon, not the other way around, in both cases Denon should be capitalized, not all caps, and it is not good that there are two styles of messages. Proper labels would be 'HEOS by Denon', 'Denon HEOS' or 'HEOS'. The latter is used as it is simpler and the manufacturer also refers to the system as HEOS.