Execution progress and result text not visible
If clicked in Hardware on the "Has Node Failed" button no text is displayed in the popup during execution and when status of request is display. This is caused by this color #FFF (which is the same as the background). 
When set to #000 the text in the popup is visible during execution of the command and when result is displayed.