Re-read back Daikin current information 2 seconds after a set.
The Daikin local API seems to be a bit hit and miss on some
controllers, meaning updates are not applied reliably. In theory,
the hardware controller should verify changes and retry, but
for now this brings forward a data poll to 2 seconds after any
update - at least that way we can quickly see if the update was
successful.