1Wire temperature sensor DeviceIDs modified to use all six bytes instead of just two.
Bytes are now XORed to create 16bit ID.

When scanning the devices, "old style" IDs are corrected in database.

NOTE: This means that if you go back from this version to regular Domoticz, new devices will be
created with the old style IDs..

BatteryLevel field in new-style temperature sensors is set to 100 instead of 255.

This allows us to detect which devices are already converted, preventing
double conversion.