Add timesync check script to domoticz.sh
This adds a script in the do_start() function to wait 30 seconds, or less until the file /run/systemd/timesync/synchronized is found indicating a successful time sync before starting Domoticz. This is to prevent systems without a real time clock to finish synchronizing to a time server, because Domoticz can crash when the time change is larger than a set watchdog timer. The script also checks if systemd-timesyncd is active and running before executing, to avoid running it on systems that do not use it. The script will always allow Domoticz to start, even if the sync fails, and all events generate a message for the user.