MQTT: Fix precision of kWh sensors
Electricity meter readings tend to have large values (hundreds of thousands
of kWh over the lifetime of the meter), but we often only care about the
consumption *today*, in units of 0.001kWh.

We lose precision in downcasting to 'float', especially after converting to
Wh. For example, a reading of 48626.531 kWh gets converted to 48626531 Wh
but...

(gdb) p 48626531.000000
$1 = 48626531
(gdb) p (float)48626531.000000
$2 = 48626532

Fix this by just using 'double' throughout the calculations.