Fix MQTT AutoDiscover device vs. sensor name confusion
In MQTT AutoDiscovery, many sensors can be attached to a single "device".
The device doesn't have an MQTT topic of its own; it is reported as a
property of each sensor, e.g.:

   "dev" : {
      "cns" : [
         [
            "mac",
            "782184bb134c"
         ]
      ],
      "ids" : "782184bb134c",
      "mdl" : "esp32-gateway",
      "mf" : "Espressif",
      "name" : "bathroom",
      "sw" : "2024.9.0-dev (Sep  6 2024, 22:54:39)"
   },

During sensor discovery, we look for an existing device with the same 'ids'
and use that if it exists, otherwise we create a new one. We keep track of
which sensors are attached to which device, for matching up things like
electricity power/usage and temp/hum/baro sensors.

However, as we discover each sensor and work out what its name should be,
we *overwrite* the name of the device it's attached to. So the device ends
up bearing the name of the last sensor that happens to be discovered on it.

Clean that up, to set the device name using only information about the
device itself, and then set the sensor name as before.

This shouldn't make any difference to the sensor names, *except* for the
combined Temperature/Humidity/Barometer devices which are later synthesised
based on the device name. That's the only time the device name actually
gets used.

Fixes: 6164