Fix MQTT AutoDiscover device names
In HA MQTT AutoDiscovery, many sensors can be attached to a single device.
The device doesn't have an MQTT topic of its own; it is reported as a
property of each sensor, e.g.:

   "dev" : {
      "cns" : [
         [
            "mac",
            "782184bb134c"
         ]
      ],
      "ids" : "782184bb134c",
      "mdl" : "esp32-gateway",
      "mf" : "Espressif",
      "name" : "bathroom",
      "sw" : "2024.9.0-dev (Sep  6 2024, 22:54:39)"
   },

During sensor discovery, we look for an existing device with the same 'ids'
and use that if it exists, otherwise we create a new one. We keep track of
which sensors are attached to which device, for matching up things like
electricity power/usage and temp/hum/baro sensors.

However, as we discover each sensor, we currently *overwrite* the device
name to include the sensor name, then blindly set the new sensor name to
the current device name. Leaving the device with the name of whatever
sensor happens to have been discovered last.

Clean that up, to set the device name using only information about the
device itself, and then set the sensor name as before but using its own
variable instead of messing with the device name.

Now my combined Temp/Hum sensors won't get such strange names.

Fixes: 6164