Add socketpair support to StoppableTask for efficient event-driven I/O
Worker threads need to wake up quickly and respond to three event types:
 - When work comes in on a socket or serial port
 - When they're asked to stop
 - When a certain time is reached (for periodic events, heartbeats, etc.)

Domoticz currently doesn't provide a way for threads to act responsively
on all three. Mostly we use StoppableTask::IsStopRequested() which will
wait for a period of time and return immediately if the thread is asked
to stop (which isn't the common case we should be optimising for). And
then we process the actual work of the thread mostly as an afterthought,
often up to a second after it actually arrived.

The alternative approach, like in MainWorker::Do_Work_On_Rx_Messages(),
is to prioritise the *work* instead. It mostly spends its time sitting in
m_rxMessageQueue.timed_wait_and_pop(), waking every five seconds just to
check if it should stop with an IsStopRequested(0) call.

The current StoppableTask implementation means that either the work, or
the stop check, is done as an 'afterthought' and not handled promptly.
And it means that threads are waking up often when they don't need to,
as well as not waking up as fast as they should when they *do*.

This change to StoppableTask makes it possible for a thread to react
quickly to all three. It changes the implementation to use a socketpair
which means that we can add a GetStopFd() method, and then a worker
thread can use a standard select() loop waiting on that fd as well as
its file descriptors for real I/O, with whatever timeout it desires.