Simplify and fix code flow in MainWorker::SetSetPointInt()
There are some hardware types we call directly into their SetSetpoint()
method, and for the rest we call WriteToHardware().

But pTypeRadiator1 is handled differently, and uses a RADIATOR1 struct
instead of the _tSetpoint. Where we call WriteToHardware(), we do the
right thing and use the RADIATOR1 struct both for that, and for the
subsequent PushAndWaitRxMessage() call.

But on a hardware type where we call SetSetpoint() directly, we fall
through to the PushAndWaitRxMessage() at the very end of the function,
which sends the wrong type of message for a pTypeRadiator1.

Switch it around just to set a bWriteToHardware flag for hardware that
isn't specially handled, then we build either the _tSetpoint or the
RADIATOR1 struct, and use the appropriate struct for both.